Moved initial action of cms routing into controller module
